repository cleaning
