+++ app/views/components/mainMenu.vue
