+++ package.json
@@ -43,12 +43,16 @@
     "bluebird": "^3.4.7",
     "body-parser": "^1.16.1",
     "config": "^1.25.1",
+    "cookie-parser": "^1.4.3",
     "ejs": "^2.5.6",
     "express": "^4.14.1",
+    "express-session": "^1.15.3",
     "express-vue": "^3.14.3",
     "jsonwebtoken": "^7.3.0",
     "mongoose": "^4.8.4",
-    "mongoose-shortid-nodeps": "^0.6.5"
+    "mongoose-shortid-nodeps": "^0.6.5",
+    "passport": "^0.3.2",
+    "passport-local": "^1.0.0"
   },
   "devDependencies": {
     "coveralls": "^2.12.0",
