+++ server.js
@@ -1,4 +1,3 @@
-let app = require('./app/app')
-
+let app = require('./app')
 app.connectDataBase();
 app.startServer();
